add foundation

+++ lyricsview/src/main/java/com/skateboard/lyricsview/LyricsItem.kt
@@ -1,3 +1,3 @@
 package com.skateboard.lyricsview
 
-data class LyricsItem(var ti:String="",var ar:String="",var al:String="",var by:String="",var offset:Int=0,var start:Int=0,var duration:Int=0,var lyrics:String="")
+data class LyricsItem(var ti:String="",var ar:String="",var al:String="",var by:String="",var offset:Long=0,var start:Long=0,var duration:Long=0,var lyrics:String="")
